Fix: handle pointers on PS2 EE
The PS2 EE GP registers are 128 bits, but the memory accesses are only
32 bits. We need to truncate the GP registers before using them in
effective memory addresses.

+++ src/Core/Expressions/BigConstant.cs
@@ -157,7 +157,7 @@ namespace Reko.Core.Expressions
         public override ushort ToUInt16() => (ushort) Value;
 
         /// <inheritdoc/>
-        public override uint ToUInt32() => (uint) Value;
+        public override uint ToUInt32() => (uint) (int) Value;
 
         /// <inheritdoc/>
         public override ulong ToUInt64() => (ulong) Value;
